ingestion-docker/requirements.txt
snowflake-connector-python
snowflake
pandas
numpy
requests
datetime
yfinance
python-dotenv
